feat: sn (#16)
* introducing messaging from the PR on 1dec

* forces messages introduced and tested in the messaging_test

* feat: social network

* feat: add ping event and social event weights

* fix: test

* clean: reorg

* merge

* fix

* clean

---------
