+++ .gitignore
@@ -6,3 +6,5 @@ game_dump.json
 Test.py
 *.pyc
 venv
+**.txt
+
